moving grid stuff to section component
