+++ src/components/NewsGroup.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { DateNewsCard, Button } from '@baltimorecounty/dotgov-components';
+import { DateNewsCard } from '@baltimorecounty/dotgov-components';
 const NewsGroup = (props) => {
 	const { Newsitems = [] } = props;
 	return (
